Added CultureObject to CulturePbjectHall.. Refactoring .

+++ mrs/mrs/Infrastructure/Data/EFRepository.cs
@@ -32,7 +32,7 @@
         /// </summary>
         /// <param name="id">The identifier.</param>
         /// <returns></returns>
-        public virtual T GetById(int id)
+        public virtual T GetById(long id)
         {
             return _dbContext.Set<T>().Find(id);
         }
@@ -50,7 +50,7 @@
         /// </summary>
         /// <param name="id">The identifier.</param>
         /// <returns></returns>
-        public virtual async Task<T> GetByIdAsync(int id)
+        public virtual async Task<T> GetByIdAsync(long id)
         {
             return await _dbContext.Set<T>().FindAsync(id);
         }
